code/requirements.txt
Flask
cassandra-driver
pandas